Add rest of user commands

+++ app/Commands/ResourceCommand.php
@@ -7,10 +7,18 @@ use App\Contracts\ResourceCommandContract;
 use App\Models\Setting;
 use Illuminate\Console\Command;
 use Illuminate\Support\Facades\File;
+use Illuminate\Support\Str;
+use Phar;
+use Symfony\Component\HttpFoundation\Request;
 use function Termwind\render;
 
 abstract class ResourceCommand extends Command implements ResourceCommandContract
 {
+    /**
+     * @var string
+     */
+    protected $method = 'GET';
+
     /**
      * Execute the console command.
      *
@@ -18,7 +26,7 @@ abstract class ResourceCommand extends Command implements ResourceCommandContrac
      */
     public function handle()
     {
-        if (! File::exists($_SERVER['HOME'].'/.perscom/database/database.sqlite')) {
+        if (Phar::running() && ! File::exists($_SERVER['HOME'].'/.perscom/database/database.sqlite')) {
             $this->error('Unable to find the local database. Please run the install command to perform the PERSCOM CLI setup.');
 
             return Command::FAILURE;
@@ -32,14 +40,50 @@ abstract class ResourceCommand extends Command implements ResourceCommandContrac
 
         $apiService = app()->make(PerscomApiServiceContract::class);
 
-        $response = $apiService->api(optional($this->option('id'), function ($id) {
+        $id = null;
+        if ($this->hasArgument('id')) {
+            $id = $this->argument('id');
+        }
+
+        $body = null;
+        if ($this->hasOption('body')) {
+            $body = json_decode($this->option('body'), true);
+
+            if (is_null($body)) {
+                $this->error('The body provided is not proper JSON. Please try again.');
+
+                return Command::FAILURE;
+            }
+        }
+
+        $response = $apiService->api(optional($id, function ($id) {
             return "$this->endpoint/$id";
-        }) ?: $this->endpoint);
+        }) ?: $this->endpoint, $this->method, $body);
 
         if (! $response->successful()) {
             $this->error($response->json('error.message', 'There was an error with your last request. Please try again.'));
 
             return Command::FAILURE;
+        }
+
+        if ($this->method === Request::METHOD_POST) {
+            $resource = Str::lower(Str::singular($this->endpoint));
+
+            $this->info("The $resource has been successfully created.");
+        }
+
+        if ($this->method === Request::METHOD_PUT) {
+            $resource = Str::lower(Str::singular($this->endpoint));
+
+            $this->info("The $resource has been successfully updated.");
+        }
+
+        if ($this->method === Request::METHOD_DELETE) {
+            $resource = Str::lower(Str::singular($this->endpoint));
+
+            $this->info("The $resource has been successfully deleted.");
+
+            return Command::SUCCESS;
         }
 
         $transformer = app()->make($this->transformer);
